Replace __float2half_rn with __float2half

+++ Eigen/src/Core/arch/CUDA/PacketMathHalf.h
@@ -231,7 +231,7 @@ template<> __device__ EIGEN_STRONG_INLINE Eigen::half predux<half2>(const half2&
 #else
   float a1 = __low2float(a);
   float a2 = __high2float(a);
-  return Eigen::half(__float2half_rn(a1 + a2));
+  return Eigen::half(__float2half(a1 + a2));
 #endif
 }
 
@@ -265,7 +265,7 @@ template<> __device__ EIGEN_STRONG_INLINE Eigen::half predux_mul<half2>(const ha
 #else
   float a1 = __low2float(a);
   float a2 = __high2float(a);
-  return Eigen::half(__float2half_rn(a1 * a2));
+  return Eigen::half(__float2half(a1 * a2));
 #endif
 }
